[modmail] #13 move exceptions to correct package
