+++ src/main/java/com/jvmrally/lambda/modmail/exception/NoSuchCategoryException.java
@@ -1,4 +1,4 @@
-package com.jvmrally.lambda.modmail;
+package com.jvmrally.lambda.modmail.exception;
 
 /**
  * NoSuchCategoryException
